[WASI] Move files from `wasi/nn` to `wasi_nn`
