+++ lib/host/CMakeLists.txt
@@ -3,4 +3,4 @@
 
 add_subdirectory(wasi)
 add_subdirectory(wasmedge_process)
-add_subdirectory(wasi/nn)
+add_subdirectory(wasi_nn)
